Edit /Post / Delete updates the rendered data
