+++ client/src/components/AllFeedback/index.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import FeedbackForm from "../FeedbackForm";
 import PreviousFeedback from "../PreviousFeedback";
 import PropTypes from "prop-types";
@@ -7,29 +7,25 @@ const AllFeedback = ({ student }) => {
 
 	const [allFeedback, setAllFeedback] = useState(student.allFeedback);
 
-	//Call data after feedback submitted
-	// const updateFeedback = () => {
-	// 	fetch(`/api/students/${student._id}`)
-	// 		.then((res) => res.json())
-	// 		.then((student) => setAllFeedback(student.allFeedback.reverse()))
-	// 		.catch((err) => console.log(err));
-	// };
-
-	//Render the page on first load
-	useEffect(()=>{
-
-		setAllFeedback(student.allFeedback);
-	}, [student]);
+	// Call data after feedback submitted
+	const updateFeedback = () => {
+		fetch(`/api/students/${student._id}`)
+			.then((res) => res.json())
+			.then((student) => {
+				setAllFeedback(student.allFeedback);
+			})
+			.catch((err) => console.log(err));
+	};
 
 	return (
 		<div>
 			<FeedbackForm
-			// updateFeedback={updateFeedback}
+				updateFeedback={updateFeedback}
 				student={student} />
 			<PreviousFeedback
 				student={student}
 				allFeedback={allFeedback}
-				// updateFeedback={updateFeedback}
+				updateFeedback={updateFeedback}
 			/>
 		</div>
 	);
